+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/adapter/HandlerAdapter.java
@@ -1,7 +1,8 @@
-package webmvc.org.springframework.web.servlet;
+package webmvc.org.springframework.web.servlet.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import webmvc.org.springframework.web.servlet.ModelAndView;
 
 public interface HandlerAdapter {
