+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/handlermapping/HandlerMappings.java
@@ -1,7 +1,6 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.handlermapping;
 
 import jakarta.servlet.http.HttpServletRequest;
-import webmvc.org.springframework.web.servlet.HandlerMapping;
 
 import java.util.HashSet;
 import java.util.Objects;
