+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JsonView.java
@@ -2,7 +2,6 @@ package webmvc.org.springframework.web.servlet.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.View;
 
 import java.util.Map;
